Initially implement ADC1 interrupt handler
This is taken from:

    STM32Cube_FW_L4_V1.17.0/Projects/NUCLEO-L496ZG/Examples_LL/ADC/ADC_SingleConversion_TriggerSW_IT

I changed the code from the interupt handler a little bit, to expect End
Of Sequence flag to be set, since this is how I configured the SW via
CubeMX.

There is a global static function handler added, called
adc1_interrupt_handler, that can be used by the singleton to run some
custom code. The custom code is called from the interrupt if it is
initialized.