Handle new samples from thread
An Active object is used when new samples arrive. The
adc1_interrupt_handler is called to push new raw sample to the message
pump. Then the Active object will call the handler passed to it when
constructing. The handler calls handle_new_sample().

This makes the ISR handler possibly short and all the other handling is
done from the task context.