Handle samples in batches
This introduces a little bit of coupling. Because of that the samples
come quite frequently, we perform such scenario:

1. ADC1 interrupt sends a sample to a queue.
2. Queue is unlocked so a context switch may happen.
3. Corresponding handler is called for a new sample.

For each new sample a context switch may happen, what takes a lots of
time. This means that we have to decrease the number of context
switches. This is done by supplying the samples to the handler in
batches. I guess we can use ADC with DMA for that.

By introducing coupling I mean that the interface is forced to be
defined basing on the details. On the other hand it is OK to define such
an interface, because it is natural to handle samples in batches.