Invoke batch handler on DMA interrupt
This turns from using ADC interrupt on single sample received to use DMA
interrupt, which is invoked when a batch of samples is sampled.

The DMA is configured to store samples into internal raw_sample_buffer.
The DMA interrupt is configured to trigger when the entire buffer is
filled. Then send_new_batch_of_samples_to_active() is invoked from the
interrupt handler. The interrupt handler can access this private method
because there is a singleton_pointer. singleton_pointer is a pointer to
the instance created with a constructor. Only one instance of
ThreadedAudioSampler can be created at a time. The DMA interrupt handler
is friend to the ThreadedAudioSampler, so it can access private methods.

send_new_batch_of_samples_to_active() sends the sampled batch of raw
samples to the Active. Active then performs conversion from raw samples
to samples from within the int range. If the external handler is
registered it is invoked from the Active context.