Switch to custom event-based synchronization
The ThreadedAudioSampler used jungles::freertos::active before. This is
wrong, because it used FreeRTOS ISR API to quit the thread.

Active is good when using messaging between multiple threads, or when
deputing job for later and splitting the tasks. Here we use DMA with
specific timely requirements. Active introduces lots of overhead for
this solution. Custom Event-based approach is much better here.

The new approach works in such a way:

1. DMA will trigger interrupt on half transfer and full transfer.
2. The raw buffer is twice the size of the converted sample buffer.
3. When half transfer IT occurs the first half of the raw sample
buffer is taken, then it is converted and propagated to the handler
registered with set_on_batch_of_samples_received_handler().
4. When full transfer IT occurs then the second half of the buffer is
converted and forwarded to the handler.
5. The half transfer and full transfer events are raised from the
interrupt using FreeRTOS Event Bits/Groups.
6. In the thread code the Events are handled -> the corresponding half
of the raw sample buffer is converted and forwarded to the external
handler.
7. Quitting the thread is synchronized through the same FreeRTOS
Event Bus - it is used to quit the endless while loop. Then the actual
termination is informed through the thread_finished_semaphore, to
prevent from destroying the Event Group before the thread has finished.