Remove non-testable test
We can't really test the stability of the sampling frequency, because
we should test here whether the clock is stable.

The samples come for sure not in orderly matter because we use RTOS,
which may affect timings.

This doesn't mean that the samples received have jitter because of
RTOS. We simply perform calculation non-deterministically.