Fix batch of samples to be std::vector<int>
This fixes the AudioSampler Interface to expect any type in the handler.
This is much better because the type used may be anything: a pair of
iterators, a const reference to some container.

We use std::vector<int> passed to the handler by value.

This is much better, because we can easily move around the vector.