Delay before disabling
This is needed, because in the Release build this has failed.

The error was that the test:
test_multiple_instances_can_be_run_one_after_another() failed (hanged),
because the previous instance was not properly stopped.

After the first instance was created, the second instance didn't
generate any interrupt, due to inproper disabling of ADC.