Test input sine sampling
This adds semi-automatic tests which test whether input sine wave is
properly sampled. The tests check against 100 Hz, 1 kHz, 10 kHz and 20
kHz.

The tests work in such a way that it is expected to provide the
corresponding sine wave to the input pin PA0 (A0 STM32L432KC Nucleo
pin). The amplitude shall be 1.6V, DC offset 1.6V. Then the sine wave is
compared against a pattern and cross correlation function is calculated.

At this stage the tests are quite unstable. Sometimes they show that the
correlation factor is almost 1 (or -1). But every second test it shows
much smaller value. It may depend on the function generator I use, which
is so-so stable (GW GFG-8019G). Eventually, when the frequency set on
the function generator matches the expected sine wave frequency every
second test passes.

The tests don't work in Release mode.