Switch to FFT based harmonic comparison
We abandon cross-correlation function based comparison, because it is
better for populations and sets, not good for signals. We use FFT
instead.

Because of that the Q DSP library (used for generating sine wave which
was the signal pattern) and CPP Utils library (implementing
cross-correlation function) are no longer needed.

The comparison works like that:

* Compute FTT and harmonic magnitudes.
* For the specific harmonic find an index in the FFT.
* Get power of the harmonic, by adding the magnitude of the
corresponding FFT sample and the magnitude of the FFT samples on the
left and right (get_sum_around())
* Get power of the noise, which is the sum of the magnitudes in the FFT,
minus the power of the harmonic from the bullet above, minus the DC
signal magnitude (FFT at index 0).
* Calculate SNR and check whether it has value bigger than 0.5, which is
acceptable level for a single harmonic.