Add interface and implementer for sampling timer
This will be used by both the ADC and DAC, to generate the same sampling
frequency in both modules. The timer will be have to implement an use
count, similar to "reference count", when using start() and stop(). This
is a must, because otherwise the timer could be stopped when it is still
needed by the other module.

The ADC doesn't use the interface yet, thus it is a TODO.