Implement basic audio DAC
This only implements the module to pass the tests in
test_audio_dac_basic.cpp. The sine wave generation is not yet
implemented.

The implementation is a copy of the implementation within the
ThreadedAudioSampler. The principle of operation is similar. We use DMA
to update the underlying raw buffer from the thread code. The raw buffer
is twice the size of the buffer with samples, which are obtained using
the Handler. After each half of the buffer transferred using DMA to DAC,
we ask the Handler for new samples and copy it to the just-updated half,
to give time for conversion, thus preventing from overwritting. This is
hard-real-time task, so the conversion should be quick. It should take
at most: 1/44100 * (raw buffer size) / 2 [s] = 1/44100 * 64 s = 1,45 ms
We have to take into account RTOS context switches into account, so the
time of conversion should be most probably around 1,2 ms, for a single
batch of samples (64 samples).

By 'conversion' I mean:

1. ADC raw sample to float conversion.
2. Applying effect (e.g. distortion).
3. float to DAC raw sample conversion.

For detailed explanation of the operation see the ThreadedAudioSampler
documentation (currently documented mostly within git commits).

In few words: there is a thread which is driven by events from the DMA
interrupts. Interrupts are raised on half transfer and full transfer of
the raw buffer. The thread fills the raw buffer by asking external
Handler for samples. The size of the buffer with samples is half of the
raw buffer size. This is to prevent from overwritting and provide
continous sample stream.