Initial impl of BasicWindowedDistortionWithMemory
This is Distortion implementation by windows, with memory, which stores
the previous batch of samples. It works in such a way, that for each
sample a absolute maximum is found in a window. If current sample has
i'th index and window size is W then the window spans from sample
'i - W' excluded to 'i' included. For that window a threshold is
calculated basing on the absolute maximum value. If the current sample's
absolute value is above the threshold, then it is clamped to the
[- window threshold] or [+ window threshold].

The implementation is initial because it is not fully tested and some
improvements and refactorings will be made.