Not working: Add filter cutoff frequency setter clock
This doesn't work yet. This adds new test which checks whether the class
is able to produce the clock. At this stage it is not.
I will investigate that further.

I checked the code agains the example:

    STM32Cube_FW_L4_V1.17.0/Projects/NUCLEO-L496ZG/Examples_LL/TIM/TIM_PWMOutput_Init

and it seems almost identical. The problem may lie in the SB16/SB18
jumpers on the nucleo board.