Benchmark distortion on the device
This adds test which checks how much time does it take for
BasicWindowedDistortionWithMemory to apply the effect.

Results are:

    INFO: Distortion apply() execution times: max: 572, min: 564, average: 567.019

Which means that, taking into consideration that we have window with 64
samples converted by ADC (or DAC) with 44100 Hz frequency, gives around
1450 us to sample a window. It means that we have 1450 us between each
transfer. Taking into account RTOS context switching and signal
adaptation calculations within ADC and DAC we have around 1000 us for
the guitar effect to apply the effect.

Around 570 us seems cool.

Whether the 500 us margin for RTOS context switching and signal
adaptation is a good assumption, it will be verified in coming tests.