Initially add tracing
This creates an interface which allows to trace events within the
firmware. Currently the audio chain itself uses it, but we will use it
in the ADC and DAC implementations as well.

The idea is to see how much time does each stage take within the
firmware itself, instead of checking it in separation. The OS itself
and context switching adds some overhead, so we need to take that into
calculation as well.

A RAII class has been added, which allows to capture events at the
beginning of the scope and at the end of the scope.

First implementation is dummy, so no capturing is implemented at this
stage.