Implement event_periodic_dumper
This EventPeriodicDumper implements EventTracer, in a way that the
captured events are dumped with specified period.

RingBuffer dependency has been added, because the events are stored in a
circular buffer. When the dumping period is reached all the events from
the ring buffer are dumped, thus effectively only the last events are
dumped. The number of the events dumped are equal to the RingBuffer
size, which is 256 currently.

I also updated the JunglesOsHelpers dependency to make use of
mutex::try_lock() which is implemented in the newer version.