Benchmark various containers on the device
For constructing and moving around, I was curious which of the
approaches is most time-effective:

* std::unique_ptr<std::array<unsigned, Size>>
* std::array<unsigned, Size>
* std::vector<unsigned>{Size>

And it turns out that the former is the most efficient timewise,
according to the logs from the benchmark:

    INFO: Performing construction and then move-construction benchmark, for 1024 and container size: 1024
    INFO: std::unique_ptr<std::array<unsigned, 1024>> duration: 49336
    INFO: std::array<unsigned, 1024> duration: 65370
    INFO: std::vector<unsigned> with size: 1024, duration: 50029
    INFO: BatchOfSamplesTemplate<1024> duration: 48951

I had to force optimization to be turned off, because otherwise the
loops would be optimized out, since they do nothing. It has been tested
on the Release build.

I did pre-optimization with using of BatchOfSamplesTemplate, being
std::unique_ptr<std::array< ... >>, but it turned out that it might be a
good choice.

My first thinking was that if we use unique_ptr to an array it will be
easily movable, because all the memory for the array will be
dynamically allocated. This must have been a good intuition.