Rework AudioChain to be run in main context
With run() and stop() we can run the AudioChain from the current thread,
decreasing number of threads needed to run the app. Since the
AudioSampler and the AudioDac are awaitables, they could also benefit
from that and not run their own contexts, from which they would have to
call the callbacks, as in the previous solution.

Design-wise we will decrease the number of threads by 2, because there
will be neither ThreadedAudioSampler, nor ThreadedAudioDac, since they
will raise events directly to the overriden *await_() function.