Simplify AudioChain
Now the AudioChain has seriously only one responsibility, which is
chaining the Audio signal from ADC, through GuitarEffect to DAC.
This allows to separate the problem of running and stopping to another
module. The tests are real simple now as well. We don't need to use
std::thread and any time-related operations at all.