Adapt ThreadedAudioSampler to new interface
It's name is no longer valid, because there is no thread involved, so
TODO is to rename it.
Since this commit the class is awaitable, according to the AudioChain
requirements.
The code didn't change a lot regarding the logic. The half-transfer
events and full-transfer events are transfered to the await_samples().
We could drop the quit event, since there is no thread to quit.

The tests pass, but the problem is that when they fail, the objects are
not properly destructed. This must be fixed.