Fix test hanging when failure
Because of that the Unity test framework uses `setjmp` and `longjmp` the
destructors are not called when the tests fail. This is really annoying,
because there will be some crash on the way because of that, since
something was not cleaned up, dangling references arise, and similar
issues.
Disabling this within Unity will return control to the test even when
the test fails.