Optimize distortion
To have the absolute values always sorted we can use std::multiset. The
problem is that we must keep track of the window: when the window moves
right, we have to drop the oldest sample and introduce a new one.
Multiset will keep the values sorted, but we have to know which samples
in the multiset must be dropped when the window moves.
To do that we introduce a FIFO queue, which contains iterators to the
multiset. Elements in the queue are sorted chronologically, so it means
that the first element in the queue corresponds to the oldest sample in
the window. So when we drop the first element in the queue (pop(), or
pop from front), and we put a new element to the queue, it emulates a
moving window. Since the queue contains links (iterators) to the
multiset, we know exactly which elements to drop from the multiset when
the window moves. This spares a lot of time, because we spare time on
finding the maximum, since multiset sorts the elements on each operation
performed on it (erase(), insert(), ...) and also we don't have to look
for which element to erase, because we store the links to it in the
queue.

multiset::begin() will return always the biggest value, when
std::greater is used as Compare type.