Use RingBuffer instead of queue
Since the queue operated like a ring buffer, we can optimize the type to
get rid of dynamic allocations.

Running:

    ./tests/host/jeff_host_tests [distortion_benchmark]

Gives higher values than for the primary solution, so this means that
this is less optimized than the primary solution. I will test it on the
device as well, but I think that this branch will be dropped.