Optimize distortion
The previous solution was not well optimized, because for each window a
maximum was searched over the whole window.
Now we keep track of the maximum within the window. The window moves one
sample to the right always. Thanks to it we can see if the new sample
(which jumped to the window because of the right-shift) is bigger than
the current absolute maximum. If it's true, then this sample is the new
current absolute maximum. We also check whether the current absolute
maximum didn't drop out of the window (was the max-left sample in the
window). If it did, we have to find a new maximum over the entire
window.